requirements.txt
# Core dependencies - carefully versioned for compatibility
torch>=2.0.0,<2.5.0
transformers>=4.34.0,<4.46.0  # Compatible with torch 2.x and sentence-transformers
sentence-transformers>=2.2.2,<3.0.0
datasets>=2.14.0
accelerate>=0.20.1,<0.35.0  # Compatible with transformers and torch
peft>=0.4.0,<0.13.0  # Compatible with transformers and accelerate

# Scientific computing - pinned for stability
numpy>=1.24.0,<2.0.0  # Avoid numpy 2.x compatibility issues
scikit-learn>=1.3.0,<1.6.0

# Logging and experiment tracking
wandb>=0.15.0
omegaconf>=2.3.0
tqdm>=4.65.0

# Configuration and utilities
PyYAML>=6.0
safetensors>=0.3.1  # Required by newer transformers

# Optional: for better performance and monitoring
psutil>=5.9.0

# Development dependencies (optional)
# Uncomment if you need these for development
# pytest>=7.0.0
# black>=23.0.0
# isort>=5.12.0